// src/types/card.ts
export enum CardColor {
  RED = "RED",
  BLUE = "BLUE",
  GREY = "GREY",
  GREEN = "GREEN",
  OTHER = "OTHER",
}

export enum CardType {
  BASIC = "BASIC",
  AGENT = "AGENT",
  AGORAPHOBE = "AGORAPHOBE",
  AHAB = "AHAB",
  AMBASSADOR = "AMBASSADOR",
  ANARCHIST = "ANARCHIST",
  ANGEL = "ANGEL",
  BLIND = "BLIND",
  BOMBBOT = "BOMBBOT",
  BOMBER = "BOMBER",
  BOUNCER = "BOUNCER",
  BUTLER = "BUTLER",
  CLONE = "CLONE",
  CLOWN = "CLOWN",
  CONMAN = "CONMAN",
  COYBOY = "COYBOY",
  CRIMINAL = "CRIMINAL",
  CUPID = "CUPID",
  DEALER = "DEALER",
  DECOY = "DECOY",
  DEMON = "DEMON",
  DOCTOR = "DOCTOR",
  DRBOOM = "DRBOOM",
  DRUNK = "DRUNK",
  ENFORCER = "ENFORCER",
  ENGINEER = "ENGINEER",
  ERIS = "ERIS",
  GAMBLER = "GAMBLER",
  HOTPOTATO = "HOTPOTATO",
  INVINCIBLE = "INVINCIBLE",
  IMMUNOLOGIST = "IMMUNOLOGIST",
  INTERN = "INTERN",
  JULIET = "JULIET",
  LEPRECHAUN = "LEPRECHAUN",
  MAID = "MAID",
  MARTYR = "MARTYR",
  MASTERMIND = "MASTERMIND",
  MAYOR = "MAYOR",
  MEDIC = "MEDIC",
  M16 = "M16",
  MIME = "MIME",
  MINION = "MINION",
  MISTRESS = "MISTRESS",
  MOBY = "MOBY",
  MUMMY = "MUMMY",
  NEGOTIATOR = "NEGOTIATOR",
  NUCLEARTYRANT = "NUCLEARTYRANT",
  NURSE = "NURSE",
  PAPARAZZO = "PAPARAZZO",
  PARANOID = "PARANOID",
  PRESIDENT = "PRESIDENT",
  PRESIDENTSDAUGHTER = "PRESIDENTSDAUGHTER",
  PRIVATEEYE = "PRIVATEEYE",
  PSYCHOLOGIST = "PSYCHOLOGIST",
  QUEEN = "QUEEN",
  RIVAL = "RIVAL",
  ROBOT = "ROBOT",
  ROMEO = "ROMEO",
  SECURITY = "SECURITY",
  SHYGUY = "SHYGUY",
  SNIPER = "SNIPER",
  BLUESPY = "BLUESPY",
  REDSPY = "REDSPY",
  SURVIVOR = "SURVIVOR",
  TARGET = "TARGET",
  THUG = "THUG",
  TINKERER = "TINKERER",
  TRAVELER = "TRAVELER",
  TUESDAYKNIGHT = "TUESDAYKNIGHT",
  USURPER = "USURPER",
  VICTIM = "VICTIM",
  WIFE = "WIFE",
  ZOMBIE = "ZOMBIE",
}

export enum CardId {
  AGENT_BLUE,
  AGENT_RED,
  AGORAPHOBE,
  AHAB,
  AMBASSADOR_BLUE,
  AMBASSADOR_RED,
  ANARCHIST,
  ANGEL_BLUE,
  ANGEL_RED,
  BLIND_BLUE,
  BLIND_RED,
  BLUETEAM,
  REDTEAM,
  BOMBBOT,
  BOMBER,
  BOUNCER_BLUE,
  BOUNCER_RED,
  BUTLER,
  CLONE,
  CLOWN_BLUE,
  CLOWN_RED,
  CONMAN_BLUE,
  CONMAN_RED,
  COYBOY_BLUE,
  COYBOY_RED,
  CRIMINAL_BLUE,
  CRIMINAL_RED,
  CUPID,
  DEALER_BLUE,
  DEALER_RED,
  DECOY,
  DEMON_BLUE,
  DEMON_RED,
  DOCTOR,
  DRBOOM,
  DRUNK,
  ENFORCER_BLUE,
  ENFORCER_RED,
  ENGINEER,
  ERIS,
  GAMBLER,
  HOTPOTATO,
  INVINCIBLE,
  IMMUNOLOGIST,
  INTERN,
  JULIET,
  LEPRECHAUN,
  MAID,
  MARTYR,
  MASTERMIND,
  MAYOR_BLUE,
  MAYOR_RED,
  MEDIC_BLUE,
  MEDIC_RED,
  M16,
  MIME_BLUE,
  MIME_RED,
  MINION,
  MISTRESS,
  MOBY,
  MUMMY_BLUE,
  MUMMY_RED,
  NEGOTIATOR_BLUE,
  NEGOTIATOR_RED,
  NUCLEARTYRANT,
  NURSE,
  PAPARAZZO_BLUE,
  PAPARAZZO_RED,
  PARANOID_BLUE,
  PARANOID_RED,
  PRESIDENT,
  PRESIDENTSDAUGHTER,
  PRIVATEEYE,
  PSYCHOLOGIST_BLUE,
  PSYCHOLOGIST_RED,
  QUEEN,
  RIVAL,
  ROBOT,
  ROMEO,
  SECURITY_BLUE,
  SECURITY_RED,
  SHYGUY_BLUE,
  SHYGUY_RED,
  SNIPER,
  REDSPY,
  BLUESPY,
  SURVIVOR,
  TARGET,
  THUG_BLUE,
  THUG_RED,
  TINKERER,
  TRAVELER,
  TUESDAYKNIGHT,
  USURPER_BLUE,
  USURPER_RED,
  VICTIM,
  WIFE,
  ZOMBIE,
}